requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiohttp-retry==2.9.1
aiosignal==1.4.0
aiosqlite==0.21.0
alembic==1.15.2
altair==5.5.0
altex==0.2.0
annotated-doc==0.0.3
annotated-types==0.7.0
anthropic==0.72.0
anyio==4.12.1
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asgiref==3.10.0
asn1crypto==1.5.1
asttokens==2.4.1
asyncpg==0.31.0
attrs==25.3.0
audioop-lts==0.2.2
backoff==2.2.1
bcrypt==4.0.1
beautifulsoup4==4.13.4
blinker==1.9.0
blis==1.3.3
boto3==1.40.11
botocore==1.40.11
brotli==1.2.0
build==1.3.0
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.11.12
cffi==1.17.1
charset-normalizer==3.4.2
chromadb==1.4.0
click==8.2.1
cloudpathlib==0.23.0
cloudpickle==3.0.0
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
confection==0.1.5
contourpy==1.3.2
cryptography==45.0.6
cycler==0.12.1
cymem==2.0.13
dataclasses-json==0.6.7
datasets==4.3.0
ddgs==9.8.0
debugpy==1.8.11
decorator==5.1.1
defusedxml==0.7.1
dill==0.4.0
distro==1.9.0
dnspython==2.8.0
docker==7.1.0
docopt==0.6.2
docstring_parser==0.17.0
docx2txt==0.9
duckduckgo_search==8.1.1
durationpy==0.10
dydantic==0.0.8
ecdsa==0.19.1
email-validator==2.3.0
entrypoints==0.4
evaluate==0.4.6
exceptiongroup==1.2.2
executing==2.1.0
faiss-cpu==1.12.0
Faker==37.5.3
fastapi==0.115.6
fastapi-cli==0.0.14
fastapi-cloud-cli==0.3.1
fastapi-utils==0.8.0
fastembed==0.7.4
fastjsonschema==2.21.1
favicon==0.7.0
filelock==3.18.0
filetype==1.2.0
flatbuffers==25.9.23
fonttools==4.59.0
fpdf==1.7.2
frozenlist==1.7.0
fsspec==2025.7.0
gitdb==4.0.12
GitPython==3.1.45
google-ai-generativelanguage==0.9.0
google-api-core==2.25.1
google-api-python-client==2.177.0
google-auth==2.48.0
google-auth-httplib2==0.2.0
google-cloud-core==2.5.0
google-cloud-modelarmor==0.3.0
google-genai==1.60.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
googletrans==4.0.2
grandalf==0.8
greenlet==3.2.3
groq==0.35.0
grpcio==1.74.0
grpcio-status==1.71.2
h11==0.16.0
h2==4.3.0
hf-xet==1.2.0
hpack==4.1.0
htbuilder==0.9.0
httpcore==1.0.9
httplib2==0.22.0
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.34.3
humanfriendly==10.0
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.5.2
iniconfig==2.3.0
ipinfo==5.3.0
ipykernel==6.29.5
ipython==8.32.0
ipython_pygments_lexers==0.1.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
kubernetes==34.1.0
langchain==1.0.3
langchain-anthropic==1.0.1
langchain-chroma==1.1.0
langchain-classic==1.0.0
langchain-community==0.4.1
langchain-core==1.2.6
langchain-experimental==0.3.4
langchain-google-community==3.0.1
langchain-google-genai==3.0.3
langchain-groq==1.0.1
langchain-huggingface==1.0.0
langchain-mcp-adapters==0.2.1
langchain-openai==1.0.1
langchain-pinecone==0.2.13
langchain-qdrant==1.1.0
langchain-text-splitters==1.0.0
langgraph==1.0.3
langgraph-checkpoint==3.0.0
langgraph-checkpoint-postgres==3.0.2
langgraph-checkpoint-sqlite==3.0.0
langgraph-prebuilt==1.0.2
langgraph-sdk==0.2.9
langmem==0.0.30
langsmith==0.4.8
llvmlite==0.46.0
loguru==0.7.3
lxml==6.0.0
Mako==1.3.10
Markdown==3.8.2
markdown-it-py==4.0.0
markdownlit==0.0.7
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.3
matplotlib-inline==0.1.7
mcp==1.25.0
mdurl==0.1.2
mem0ai==1.0.1
mmh3==5.2.0
more-itertools==10.8.0
mpmath==1.3.0
multidict==6.6.3
multiprocess==0.70.16
murmurhash==1.0.15
mypy_extensions==1.1.0
narwhals==1.48.1
nbformat==5.10.4
nest_asyncio==1.6.0
networkx==3.5
nltk==3.9.2
numba==0.63.1
numpy==2.3.2
oauthlib==3.3.1
onnxruntime==1.23.2
openai==2.6.1
opentelemetry-api==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-grpc==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
orjson==3.10.18
ormsgpack==1.11.0
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.2
pandas==2.3.1
parso==0.8.4
passlib==1.7.4
patsy==1.0.1
pdf2image==1.17.0
pickleshare==0.7.5
pillow==11.3.0
pinecone==7.3.0
pinecone-client==6.0.0
pinecone-plugin-assistant==1.8.0
pinecone-plugin-interface==0.0.7
pipreqs==0.4.13
platformdirs==4.3.6
plotly==6.2.0
plotly-express==0.4.1
pluggy==1.6.0
portalocker==3.2.0
posthog==5.4.0
preshed==3.0.12
primp==0.15.0
prometheus_client==0.22.1
prompt_toolkit==3.0.48
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
psutil==5.9.8
psycopg==3.3.2
psycopg-binary==3.3.2
psycopg-pool==3.3.0
psycopg2-binary==2.9.11
pure_eval==0.2.3
pwdlib==0.3.0
py_rust_stemmers==0.1.5
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAudio==0.2.14
pybase64==1.4.2
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydeck==0.9.1
pydub==0.25.1
Pygments==2.19.1
PyJWT==2.10.1
pymdown-extensions==10.16.1
PyMySQL==1.1.2
pyOpenSSL==25.1.0
pyparsing==3.2.3
pypdf==6.1.3
PyPDF2==3.0.1
pyperclip==1.9.0
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
PySocks==1.7.1
pytest==9.0.1
python-dateutil==2.9.0
python-docx==1.2.0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
python-pptx==1.0.2
pytube==15.0.0
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.2.0
qdrant-client==1.15.1
referencing==0.36.2
regex==2024.11.6
replicate==1.0.7
requests==2.32.5
requests-file==3.0.1
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.2.0
rich-toolkit==0.15.1
rignore==0.7.6
rpds-py==0.26.0
rsa==4.9.1
s3transfer==0.13.1
safetensors==0.5.3
scikit-learn==1.7.1
scipy==1.16.1
seaborn==0.13.2
secure==1.0.1
selenium==4.39.0
sentence-transformers==5.1.2
sentry-sdk==2.43.0
setuptools==78.1.1
shellingham==1.5.4
simsimd==6.5.3
six==1.17.0
smart_open==7.5.0
smmap==5.0.2
sniffio==1.3.1
snowflake-connector-python==3.17.1
snowflake-snowpark-python==1.36.0
socksio==1.0.0
sortedcontainers==2.4.0
soupsieve==2.7
spacy==3.8.11
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SpeechRecognition==3.14.3
SQLAlchemy==2.0.41
SQLAlchemy-Utils==0.42.0
sqlite-vec==0.1.6
sqlmodel==0.0.18
srsly==2.5.2
sse-starlette==3.1.2
st-annotated-text==4.0.2
st-theme==1.2.3
stack_data==0.6.3
standard-aifc==3.13.0
standard-chunk==3.13.0
starlette==0.41.3
statsmodels==0.14.5
streamlit==1.47.1
streamlit-avatar==0.1.3
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.2
streamlit-embedcode==0.1.2
streamlit-extras==0.7.6
streamlit-image-coordinates==0.4.0
streamlit-keyup==0.3.0
streamlit-notify==0.3.1
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
streamlit_faker==0.0.4
sympy==1.14.0
tenacity==9.1.2
thinc==8.3.10
threadpoolctl==3.6.0
tiktoken==0.12.0
tldextract==5.1.3
tokenizers==0.21.4
toml==0.10.2
tomlkit==0.13.3
torch==2.7.1
torchaudio==2.7.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.15.0
transformers==4.54.0
trio==0.32.0
trio-websocket==0.12.2
trustcall==0.0.39
typer==0.20.0
typer-slim==0.20.0
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.3.0
uuid_utils==0.12.0
uvicorn==0.38.0
validators==0.35.0
wasabi==1.1.3
watchdog==6.0.0
watchfiles==1.1.1
wcwidth==0.2.13
weasel==0.4.3
webdriver-manager==4.0.2
websocket-client==1.9.0
websockets==15.0.1
wheel==0.45.1
wikipedia==1.4.0
win32_setctime==1.2.0
wordcloud==1.9.4
wrapt==2.0.1
wsproto==1.3.2
xgboost==3.0.4
xlsxwriter==3.2.5
xxhash==3.6.0
yarg==0.1.10
yarl==1.20.1
youtube-transcript-api==1.2.1
yt-dlp==2025.7.21
zenml==0.91.2
zipp==3.21.0
zstandard==0.23.0